@@ -4,24 +4,24 @@ version: 1.0.0+1
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.17.6 <3.0.0"
+  sdk: '>=3.1.4 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  provider: ^6.0.3
-  shared_preferences: ^2.0.15
+  provider: ^6.1.1
+  shared_preferences: ^2.2.2
   hive: ^2.2.3
   hive_flutter: ^1.1.0
 
 dev_dependencies:
-  intl: ^0.17.0
+  intl: ^0.18.1
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.1
+  flutter_lints: ^3.0.1
 
 flutter:
   generate: true
